requirements.txt
click==8.1.2
cssutils==2.4.0
Flask==2.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
Pillow==9.1.0
Werkzeug==2.1.1
